tfserving_fatAPI_Docker/requirements.txt
fastapi==0.70.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
numpy==1.19.5
uvicorn==0.15.0
python-multipart==0.0.5
Jinja2==3.0.2
Pillow==8.4.0
aiofiles==0.7.0
requests==2.26.0
requests-oauthlib==1.3.0

